requirements.txt
# Core LLM & Framework
ollama
litellm
python-dotenv

# Spec Parsing & IR
lark

# Sandbox & Security
RestrictedPython

# Verification & Testing
pytest
PyYAML

# UI & Utilities
rich
beautifulsoup4
lxml
markdown
